+++ CHANGELOG.md
@@ -39,6 +39,7 @@ hangeLog
 - BUGFIX: fix "t2t-summary" node for more correct counting of sentences
 - BUGFIX: fix "t2t-google" node by awaiting the close method on shutdown
 - BUGFIX: fix "xio-vban" node by awaiting all VBANServer operations
+- BUGFIX: avoid unbounded active captures memory growth in Audio Worklet
 - BUGFIX: do not throw an exception from Stream::read() operations
 - UPDATE: upgrade NPM dependencies and patches
 - CLEANUP: remove second/dead "open" WebSocket handler in "t2a-openai" node